requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
cssselect==1.2.0
distro==1.9.0
fake-useragent==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
linkify-it-py==2.0.3
lxml==5.2.2
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
openai==1.37.1
parse==1.20.2
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-html==0.10.0
rich==13.7.1
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==1.26.19
w3lib==2.2.1
websockets==10.4
zipp==3.19.2
